Add system reload tools and smart button press

+++ lib/system.py
@@ -1,13 +1,21 @@
-import pickle, os, time
+import pickle, os, time, subprocess
 
+from lib import cache
+
+MPD_STATE_FILE_NAME = '/Users/adamp/.mpd/state'
+DB_FILE_NAME = '/Users/adamp/.mpd/database'
 TIME_FILE_NAME = '/Users/adamp/.mpd/hypnotime'
-PLAYLIST_FILE_NAME = '/Users/adamp/.mpd/hypnoplaylist'
+CURRENT_PLAYLIST_FILE_NAME = '/Users/adamp/.mpd/hypnoplaylist'
 
 #
 ##
 ####
 # system utils:
+#  - getCurrentPlaylistName() returns the name of the current playlist
+#  - setCurrentPlaylistName(playlistName) set the system current playlist
 #  - getCurrentSystemTimeMs() returns system time in milliseconds
+#  - setCurrentSystemTimeMs(newSystemTime) set the system current time
+#  - setAliasedSystemTimeMs(newAliasedTimeMs) set the system time so that newAliasedTimeMs is now
 #  - reset() starts system time over at zero
 ####
 ##
@@ -15,7 +23,7 @@ PLAYLIST_FILE_NAME = '/Users/adamp/.mpd/hypnoplaylist'
 
 # Return the name of the currently playing list, or raise exception
 def getCurrentPlaylistName():
-    playlistName = readFile(PLAYLIST_FILE_NAME)
+    playlistName = readFile(CURRENT_PLAYLIST_FILE_NAME)
     if (playlistName is None):
         raise SystemError("Currently playing playlist not set")
     else:
@@ -23,11 +31,8 @@ def getCurrentPlaylistName():
 
 # Set the name of the currently playing list
 def setCurrentPlaylistName(playlistName):
-    writeFile(PLAYLIST_FILE_NAME, playlistName)
+    writeFile(CURRENT_PLAYLIST_FILE_NAME, playlistName)
 
-def newSystemTimeForAliasedStartTime(aliasedStartTimeMs):
-    intCurrentTimeMs = int(time.time() * 1000)
-    return  intCurrentTimeMs - aliasedStartTimeMs
 
 # Return the current system time in ms, AKA time since system start
 def getCurrentSystemTimeMs():
@@ -46,6 +51,11 @@ def getCurrentSystemTimeMs():
 def setCurrentSystemTimeMs(newSystemTime):
     writeFile(TIME_FILE_NAME, newSystemTime)
 
+# Set the current system time to an aliased time (AKA now - targetTime)
+def setAliasedSystemTimeMs(newAliasedTimeMs):
+    intCurrentTimeMs = int(time.time() * 1000)
+    setCurrentSystemTimeMs(intCurrentTimeMs - int(newAliasedTimeMs))
+
 # Read a file from disk that we wrote before (formatted for pickle)
 def readFile(filename):
     try:
@@ -59,11 +69,30 @@ def writeFile(filename, data):
     with open(filename, 'wb') as filehandle:
         pickle.dump(data, filehandle)
 
-#
-# TODO MORE STUFF HERE
-#
-def reset():
-    os.remove(TIME_FILE_NAME)
+# Restart MPD!  After this you will need to re-add files to the system using -l
+def restartMpd():
+    subprocess.check_output(['killall', 'mpd'])
+
+    if (os.path.isfile(DB_FILE_NAME)):
+        os.remove(DB_FILE_NAME)
+
+    if (os.path.isfile(MPD_STATE_FILE_NAME)):
+        os.remove(MPD_STATE_FILE_NAME)
+
+    subprocess.check_output(['mpd'])
+
+# Reset the system to a correct initial state
+def reset():  
+    subprocess.check_output(['mpc', 'repeat', 'on'])
+    subprocess.check_output(['mpc', 'crossfade', '1'])
+
+    if (os.path.isfile(TIME_FILE_NAME)):
+        os.remove(TIME_FILE_NAME)
+
+    if (os.path.isfile(CURRENT_PLAYLIST_FILE_NAME)):
+        os.remove(CURRENT_PLAYLIST_FILE_NAME)
+
+    cache.clearCache()
 
 # error message
 class SystemError(RuntimeError):
